refactor(test): merge topology/rbac files into canonical test modules; add PR-6372 coverage
Merge test_cluster_topology.py → test_cluster.py
- Added HetznerAPIException, create_test_virtual_machine,
  create_test_virtual_machine_image imports
- Appended _make_cluster() helper and 9 test classes:
  TestClusterValidation, TestClusterFirewallHelpers,
  TestClusterServerCreationGuards, TestHetznerProvisionFailure,
  TestNatServerSupport, TestVirtualMachineDataDiskSnapshot,
  TestVirtualMachineVolumeHelpers, TestVirtualMachineServerRouting,
  TestVirtualMachineStatusMaps
- Deleted test_cluster_topology.py

Merge test_team_rbac.py → test_team.py
- Added contextmanager, nullcontext, MagicMock, make_autoname,
  PressRole, PressRolePermission, get_team_members imports
- Appended rbac helpers (user_context, get_current_team_ctx,
  _make_press_user, _create_press_role, _add_press_role_user,
  _patch_sync_press_role) and 12 test classes
- Deleted test_team_rbac.py

Add PR-6372 coverage (previously empty stub files)
test_app_release_approval_request.py:
  - TestApprovalRequestGuards (4 tests): before_insert guards
    (update_release_status, duplicate request, yanked release,
    same-source conflict)
  - TestApprovalRequestStatusPropagation (3 tests): on_update
    propagates Approved / Rejected / Cancelled to linked release
  - TestApprovalRequestAutoApproval (3 tests): before_save
    auto-approves featured apps and auto-release teams

test_app_release_difference.py:
  - TestIsMigrateNeeded (12 tests): file-path classification
    for patches.txt, hooks.py, fixtures/, custom/, languages.json,
    doctype JSON, and safe files
  - TestAppReleaseDifferenceValidate (2 tests): same-release guard
  - TestHasBranchChanged (2 tests): branch comparison logic
