+++ .secrets.baseline
@@ -424,7 +424,7 @@
         "filename": "press/press/doctype/cluster/test_cluster.py",
         "hashed_secret": "a94a8fe5ccb19ba61c4c0873d391e987982fbbd3",
         "is_verified": false,
-        "line_number": 46
+        "line_number": 54
       }
     ],
     "press/press/doctype/invoice/fixtures/stripe_payment_intent_succeeded_webhook.json": [
@@ -524,5 +524,5 @@
       }
     ]
   },
-  "generated_at": "2026-05-20T16:41:21Z"
+  "generated_at": "2026-05-26T11:22:21Z"
 }
